+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/paging/SingleRowColumnPager.java
@@ -20,6 +20,7 @@ import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Optional;
 
 import javax.annotation.Nullable;
 
@@ -33,6 +34,7 @@ import org.apache.cassandra.thrift.SliceRange;
 import org.apache.thrift.TException;
 
 import com.google.common.collect.AbstractIterator;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.keyvalue.api.RangeRequests;
@@ -89,7 +91,7 @@ public class SingleRowColumnPager {
                 InetSocketAddress host = clientPool.getRandomHostForKey(rowKey.array());
                 try {
                     List<ColumnOrSuperColumn> cells = clientPool.runWithRetryOnHost(host, this::getPage);
-                    end = (cells.size() < pageSize);
+                    end = cells.size() < pageSize;
                     if (cells.isEmpty()) {
                         return endOfData();
                     } else {
@@ -103,29 +105,51 @@ public class SingleRowColumnPager {
         }
 
         private List<ColumnOrSuperColumn> getPage(Cassandra.Client client) throws TException {
-            SliceRange sliceRange = new SliceRange(
-                    getStartColumn(),
-                    ByteBuffer.wrap(PtBytes.EMPTY_BYTE_ARRAY),
-                    false, // reversed
-                    pageSize);
-            SlicePredicate slicePred = new SlicePredicate();
-            slicePred.setSlice_range(sliceRange);
-            return queryRunner.run(client, tableRef,
-                    () -> client.get_slice(rowKey, columnParent, slicePred, consistencyLevel));
+            Optional<ByteBuffer> startColumn = getStartColumn();
+            if (startColumn.isPresent()) {
+                SliceRange sliceRange = new SliceRange(
+                        startColumn.get(),
+                        ByteBuffer.wrap(PtBytes.EMPTY_BYTE_ARRAY),
+                        false, // reversed
+                        pageSize);
+                SlicePredicate slicePred = new SlicePredicate();
+                slicePred.setSlice_range(sliceRange);
+                return queryRunner.run(client, tableRef,
+                        () -> client.get_slice(rowKey, columnParent, slicePred, consistencyLevel));
+            } else {
+                return ImmutableList.of();
+            }
         }
 
-        private ByteBuffer getStartColumn() {
+        // Given the last seen column, compute the next start column
+        private Optional<ByteBuffer> getStartColumn() {
             if (lastSeenColumn == null) {
-                return ByteBuffer.wrap(PtBytes.EMPTY_BYTE_ARRAY);
+                // An empty byte array means "start from the beginning of the row"
+                return Optional.of(ByteBuffer.wrap(PtBytes.EMPTY_BYTE_ARRAY));
             } else {
+                // We need to "increment" the last seen (colName, ts) pair to the lexicographically next pair.
+                // Note that timestamps are stored in descending order. For example, the next pair for ("a", 5)
+                // will be ("a", 4).
                 Pair<byte[], Long> colNameAndTs = CassandraKeyValueServices.decomposeName(lastSeenColumn);
                 if (colNameAndTs.getRhSide() == Long.MIN_VALUE) {
-                    byte[] nextColName = RangeRequests.nextLexicographicName(colNameAndTs.getLhSide());
-                    return CassandraKeyValueServices.makeCompositeBuffer(nextColName, Long.MAX_VALUE);
+                    // This is an edge case that will never happen in reality, but technically we have
+                    // to handle it anyway. If the last seen timestamp equals minimum possible value of a 64-bit
+                    // signed integer, we can't decrement it, so we have to carry over to the colName.
+                    // E.g. ("a", Long.MIN_VALUE) -> ("b", Long.MAX_VALUE)
+                    byte[] nextColName = RangeRequests.getNextStartRowUnlessTerminal(false, colNameAndTs.getLhSide());
+                    if (nextColName == null) {
+                        // An even "edgier" case: in addition to reaching the minimum timestamp, we also reached
+                        // the last possible colName, so there is nowhere to increment.
+                        return Optional.empty();
+                    } else {
+                        return Optional.of(CassandraKeyValueServices.makeCompositeBuffer(nextColName, Long.MAX_VALUE));
+                    }
                 } else {
-                    return CassandraKeyValueServices.makeCompositeBuffer(
+                    // Timestamps are in descending order, so just decrement the timestamp to find the next column.
+                    // E.g., ("a", 4) -> ("a", 3)
+                    return Optional.of(CassandraKeyValueServices.makeCompositeBuffer(
                             colNameAndTs.getLhSide(),
-                            colNameAndTs.getRhSide() - 1);
+                            colNameAndTs.getRhSide() - 1));
                 }
             }
         }